requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
awscli==1.32.98
awscli-local==0.21
blinker==1.8.2
boto3==1.28.84
botocore==1.34.98
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docutils==0.16
dulwich==0.21.7
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.7
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
keyring==24.3.1
localstack-client==2.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
motor==3.4.0
msgpack==1.0.8
multidict==6.0.5
nodeenv==1.8.0
openai==0.28.0
orjson==3.10.3
packaging==24.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pre-commit==3.7.1
ptyprocess==0.7.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.2
pyproject_hooks==1.1.0
pytest==8.2.1
pytest-asyncio==0.23.7
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rapidfuzz==3.9.1
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.7.2
s3transfer==0.10.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomlkit==0.12.5
tqdm==4.66.4
trove-classifiers==2024.5.22
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.0.7
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
xattr==1.1.0
yarl==1.9.4
